[simplify] Remove `base_url` Redux data flow
As it turns out, the ages-old concept of “relative URL” works just as
well. WHO WOULDA THUNK IT??! (No pun intended)

There used to be quite a bit of passing around of the document's base
URL, so as to make URLs absolute (for XHR-fetchable things such as
twigs and documentation). In fact, since the whole reader is a
single-page Web app with React-style `/#/` URLs, the base URL for
relative links is always the same. Therefore, “naive” Twig relative
URIs (e.g. in imports) just work.

- Un-plumb all `props.navigation.base_url` / `basePath` baton passing
- Remove the corresponding Redux pipework (setters and reducers)