Improve dropdown position on small devices
The problem: when the breadcrumb item is too close to the right side of the window, the dropdown might be partially hidden outside of the viewport.

The solution: if the width of the dropdown is greater than the space between the button and the right side of the viewport, add class "open-left" to ".dropdown-menu" and use it to change the position of the dropdown.